requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
pycryptodome==3.19.1
python-dotenv==1.0.0